yarrariver: Explore Yarra River Water Quality

An installable R package with a Shiny app and documented datasets to explore Yarra River water quality (1990s vs recent).

See also

Author

Maintainer: Kaiyi Ye kyee0020@student.monash.edu